requirements.txt
langchain==0.0.262
langchain-community==0.0.11
langchain-core==0.1.0
deepseek-api==0.1.0
qdrant-client==1.5.4
python-dotenv==1.0.0
pydantic==2.4.2
tiktoken==0.5.1